custom command encoder works

+++ Autopilot/Mavlink/MavlinkFunctions.hpp
@@ -79,7 +79,7 @@ mavlink_encoding_status_t Mavlink_encoder(Message_IDs_t id, mavlink_message_t *m
 
 int test__encode_then_decode(void);
 
-int custom_mavlink_msg__begin_takeoff_command_encode(uint8_t system_id, uint8_t component_id, mavlink_message_t* message, const mavlink_custom_cmd_takeoff_t* struct_ptr);
+uint16_t custom_mavlink_msg__begin_takeoff_command_encode(uint8_t system_id, uint8_t component_id, mavlink_message_t* message, const mavlink_custom_cmd_takeoff_t* struct_ptr);
 void custom_mavlink_msg__begin_takeoff_command_decode(const mavlink_message_t* message, mavlink_custom_cmd_takeoff_t* takeoff_command);
